+++ lizard.py
@@ -221,7 +221,7 @@ def arg_parser(prog=None):
     parser.add_argument("-s", "--sort",
                         help='''Sort the warning with field. The field can be
                         nloc, cyclomatic_complexity, token_count,
-                        p#arameter_count, etc. Or an customized field.''',
+                        parameter_count, etc. Or an customized field.''',
                         action="append",
                         dest="sorting",
                         default=[])
@@ -306,7 +306,8 @@ class FunctionInfo(Nesting):  # pylint: disable=R0902
 
     @property
     def parameters(self):
-        matches = [re.search(r'(\w+)(\s=.*)?$', f) for f in self.full_parameters]
+        matches = [re.search(r'(\w+)(\s=.*)?$', f)
+                   for f in self.full_parameters]
         return [m.group(1) for m in matches if m]
 
     def add_to_function_name(self, app):