+++ lizard_languages/objc.py
@@ -1,11 +1,12 @@
 ''' Language parser for Python '''
 
+from .code_reader import CodeStateMachine
 from .clike import CLikeReader, CLikeStates, CLikeNestingStackStates
 
 
 class ObjCReader(CLikeReader):
 
-    ext = ['m']
+    ext = ['m', 'mm']
     language_names = ['objectivec', 'objective-c', 'objc']
 
     def __init__(self, context):
@@ -24,7 +25,9 @@ class ObjCReader(CLikeReader):
 class ObjCStates(CLikeStates):  # pylint: disable=R0903
     def _state_global(self, token):
         super(ObjCStates, self)._state_global(token)
-        if token == '(':
+        if token == 'typedef':
+            self.next(self._typedef, token)
+        elif token == '(':
             self.next(self._state_dec, token)
 
     def _state_dec_to_imp(self, token):
@@ -64,3 +67,7 @@ class ObjCStates(CLikeStates):  # pylint: disable=R0903
 
     def _state_objc_param(self, _):
         self._state = self._state_objc_dec
+
+    @CodeStateMachine.read_until_then(';')
+    def _typedef(self, _, tokens):
+        self.next(self._state_global)