+++ test/test_languages/testCAndCPP.py
@@ -98,6 +98,11 @@ class Test_c_cpp_lizard(unittest.TestCase):
         self.assertEqual(0, result[0].parameter_count)
 
     def test_function_with_1_param(self):
+        result = get_cpp_function_list("int fun(aa bb){}")
+        self.assertEqual(1, result[0].parameter_count)
+        self.assertEqual(["bb"], result[0].parameters)
+
+    def test_function_with_1_ref_param(self):
         result = get_cpp_function_list("int fun(aa * bb){}")
         self.assertEqual(1, result[0].parameter_count)
         self.assertEqual(["bb"], result[0].parameters)
@@ -201,15 +206,15 @@ class Test_c_cpp_lizard(unittest.TestCase):
         result = get_cpp_function_list("template<> class c<double> {"
                                        "void f() {}};")
         self.assertEqual(1, len(result))
-        self.assertEqual("c<double>::f", result[0].name)
+        self.assertEqual("c::f", result[0].name)
         result = get_cpp_function_list("template<>"
                                        "class c<double, int> {void f() {}};")
         self.assertEqual(1, len(result))
-        self.assertEqual("c<double,int>::f", result[0].name)
+        self.assertEqual("c::f", result[0].name)
         result = get_cpp_function_list("namespace ns { template<>"
                                        "class c<double> {void f() {}}; }")
         self.assertEqual(1, len(result))
-        self.assertEqual("ns::c<double>::f", result[0].name)
+        self.assertEqual("ns::c::f", result[0].name)
 
     def test_template_class_partial_specialization(self):
         result = get_cpp_function_list(
@@ -220,19 +225,19 @@ class Test_c_cpp_lizard(unittest.TestCase):
         result = get_cpp_function_list("template<typename T> class c<int,T> {"
                                        "void f(T t) {}};")
         self.assertEqual(1, len(result))
-        self.assertEqual("c<int,T>::f", result[0].name)
+        self.assertEqual("c::f", result[0].name)
         result = get_cpp_function_list("template<class T> class c<int,T> {"
                                        "void f(T t) {}};")
         self.assertEqual(1, len(result))
-        self.assertEqual("c<int,T>::f", result[0].name)
+        self.assertEqual("c::f", result[0].name)
         result = get_cpp_function_list("template<class T, typename S>"
                                        "class c<int,T,S> {void f(T t) {}};")
         self.assertEqual(1, len(result))
-        self.assertEqual("c<int,T,S>::f", result[0].name)
+        self.assertEqual("c::f", result[0].name)
         result = get_cpp_function_list("namespace ns { template<class T>"
                                        "class c<int,T> {void f(T t) {}}; }")
         self.assertEqual(1, len(result))
-        self.assertEqual("ns::c<int,T>::f", result[0].name)
+        self.assertEqual("ns::c::f", result[0].name)
 
     def test_template_function(self):
         result = get_cpp_function_list("template<typename T> void f(T t) {}")
@@ -632,7 +637,6 @@ class Test_Big(unittest.TestCase):
         result = get_cpp_function_list(code)
         self.assertEqual(1, len(result))
         self.assertEqual(3, result[0].cyclomatic_complexity)
-
 
 
 class Test_Dialects(unittest.TestCase):